// src/api/openapi/models/NotificationSettingsDataRequestDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NotificationSettingsNewTripDataRequestDto } from './NotificationSettingsNewTripDataRequestDto';
import {
    NotificationSettingsNewTripDataRequestDtoFromJSON,
    NotificationSettingsNewTripDataRequestDtoFromJSONTyped,
    NotificationSettingsNewTripDataRequestDtoToJSON,
    NotificationSettingsNewTripDataRequestDtoToJSONTyped,
} from './NotificationSettingsNewTripDataRequestDto';

/**
 * 
 * @export
 * @interface NotificationSettingsDataRequestDto
 */
export interface NotificationSettingsDataRequestDto {
    /**
     * 
     * @type {NotificationSettingsNewTripDataRequestDto}
     * @memberof NotificationSettingsDataRequestDto
     */
    newTrip?: NotificationSettingsNewTripDataRequestDto;
}

/**
 * Check if a given object implements the NotificationSettingsDataRequestDto interface.
 */
export function instanceOfNotificationSettingsDataRequestDto(value: object): value is NotificationSettingsDataRequestDto {
    return true;
}

export function NotificationSettingsDataRequestDtoFromJSON(json: any): NotificationSettingsDataRequestDto {
    return NotificationSettingsDataRequestDtoFromJSONTyped(json, false);
}

export function NotificationSettingsDataRequestDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationSettingsDataRequestDto {
    if (json == null) {
        return json;
    }
    return {
        
        'newTrip': json['newTrip'] == null ? undefined : NotificationSettingsNewTripDataRequestDtoFromJSON(json['newTrip']),
    };
}

export function NotificationSettingsDataRequestDtoToJSON(json: any): NotificationSettingsDataRequestDto {
    return NotificationSettingsDataRequestDtoToJSONTyped(json, false);
}

export function NotificationSettingsDataRequestDtoToJSONTyped(value?: NotificationSettingsDataRequestDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'newTrip': NotificationSettingsNewTripDataRequestDtoToJSON(value['newTrip']),
    };
}