// src/api/openapi/models/NotificationSettingsDataResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { NotificationSettingsNewTripDataResponseDto } from './NotificationSettingsNewTripDataResponseDto';
import {
    NotificationSettingsNewTripDataResponseDtoFromJSON,
    NotificationSettingsNewTripDataResponseDtoFromJSONTyped,
    NotificationSettingsNewTripDataResponseDtoToJSON,
    NotificationSettingsNewTripDataResponseDtoToJSONTyped,
} from './NotificationSettingsNewTripDataResponseDto';

/**
 * 
 * @export
 * @interface NotificationSettingsDataResponseDto
 */
export interface NotificationSettingsDataResponseDto {
    /**
     * 
     * @type {NotificationSettingsNewTripDataResponseDto}
     * @memberof NotificationSettingsDataResponseDto
     */
    newTrip?: NotificationSettingsNewTripDataResponseDto;
}

/**
 * Check if a given object implements the NotificationSettingsDataResponseDto interface.
 */
export function instanceOfNotificationSettingsDataResponseDto(value: object): value is NotificationSettingsDataResponseDto {
    return true;
}

export function NotificationSettingsDataResponseDtoFromJSON(json: any): NotificationSettingsDataResponseDto {
    return NotificationSettingsDataResponseDtoFromJSONTyped(json, false);
}

export function NotificationSettingsDataResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationSettingsDataResponseDto {
    if (json == null) {
        return json;
    }
    return {
        
        'newTrip': json['newTrip'] == null ? undefined : NotificationSettingsNewTripDataResponseDtoFromJSON(json['newTrip']),
    };
}

export function NotificationSettingsDataResponseDtoToJSON(json: any): NotificationSettingsDataResponseDto {
    return NotificationSettingsDataResponseDtoToJSONTyped(json, false);
}

export function NotificationSettingsDataResponseDtoToJSONTyped(value?: NotificationSettingsDataResponseDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'newTrip': NotificationSettingsNewTripDataResponseDtoToJSON(value['newTrip']),
    };
}